Change data type of PK in StatusPage struct
Changed the data type of PK from int64 to int in the StatusPage struct. This makes it more consistent with how we are using integers elsewhere in our codebase.

+++ pkg/upapi/ep_statuspages.go
@@ -3,7 +3,7 @@ package upapi
 import "context"
 
 type StatusPage struct {
-	PK                        int64  `json:"pk,omitempty"`
+	PK                        int    `json:"pk,omitempty"`
 	URL                       string `json:"url,omitempty"`
 	Name                      string `json:"name"`
 	VisibilityLevel           string `json:"visibility_level"`
